api/requirements.txt
certifi==2024.2.2
chardet==5.2.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==5.0.2
django-braces==1.15.0
django-filter==23.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
openapi-codec==1.3.2
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
uritemplate==4.1.1
urllib3==2.2.0
config==0.5.1
coverage==7.4.1
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework-jwt-5==1.13.0
drf-writable-nested==0.7.0
oauthlib==3.2.2
pbr==6.0.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
stevedore==5.1.0
virtualenv==20.25.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
virtualenvwrapper-win==1.2.7
whitenoise==6.6.0
setuptools==69.0.3
pillow==10.2.0
